move row clicked actions into actions.js
Conflicts:
	packages/perspective-viewer/src/js/view.js
